Flashing via BDM now reasonably fast again
I rewrote the RAM flash routine (again!), keeping the beginning the same, and
adding some crazy loop-unrolled madness at the end for the use of the BDM
code.

Tickled a bug in my BDM USB firmware, where zero-length writes were not
queuing up a status IN request.

While not *lightning* fast, it's fast enough that it's not embarrassing.

Maybe now I can get back to work on the AVR code... ;-)