BDM and AVR prog code uses RAM more efficiently
This is a pretty minor checkin, but I think it solves two small problems that
were bothering me: having special BDM flashing code taking up space in the
image, and needing a separate variable space for "system" variables (as
opposed to application variables).

I had previously created separate spaces for vars, zvars, and xvars, but I
think I was using them incorrectly.

I created xvars so that "system" code (mostly the serial and USB chat code,
but also the BDM flash routine) could have a place to store a couple of bytes
of variables (in zero page) necessary to it working properly. I wanted to keep
them out of the way of "application" code (which could start allocating zpage
code or variables at the beginning of RAM, @ram), so xvars are allocated in
the zero page top-down from 0100.

For a while now I've allocated the BDM-over-USB and AVRprog-over-USB variables
as xvars, but I realised, finally, that that was a mistake.

So, the changes in this commit are:

First, I changed the variables used by the BDM-over-USB and AVR prog over USB
code to be zvars (allocated from the bottom of the zero page) instead of
xvars.

Currently the only xvars are required for executing of the RAM-based flash
programming routine.

Second, I changed the BDM flash routine so that even though it's compiled into
the image (so it can be copied to BDM targets) the extra space for the
BDM-specific parts are "unallocated", and RAM variables can be allocated
there.