Refactored AVR programming code
We load an interaction layer that talks to firmware over serial or USB,
and then on top of that we load the high-level programming code.

I'm still not completely sure of some of the factorings, and this code
is *untested*!

I also need to figure out whether to implement bulk reads and writes in
the serial firmware, or to fake it on the host.