The beautiful beginnings of an AVR disassembler!
Actually, it does quite a lot already, and I'm very happy with it. Structurally
it is very simple, and looks just like the table in the file
Instruction_formats (not coincidentally).

I didn't worry at all about making it efficient. I wanted it to be, if
possible, both pretty and simple.

What it can do is print out the meaning of a single-word instruction. What it
can't do is print out multi-word instrs, or generate an effective address (for
fancy interactive disassembly a la ARM).

Those things come next.

Maybe.