Added bulk read and write to AVR prog over serial
It was as hopelessly slow as the USB code was without this.

I also added a curious change: when starting a sub-command-loop (like
for AVR prog, or BDM, or whatever) I'd like the firmware to announce
itself, and the the host to check that it's talking to something that it
recognises. This change adds an implementation of that for the avrprog
over serial code.