Beginnings of AVR chat (via IIC to chip) over serial
Currently not working, since there is nothing to talk to (yet) on the
other end (ie, AVR firmware), and I was having issues getting the IIC on
the S08 to do what I wanted. It seemed like I was always reading back
the last byte written, instead of FF (which is what I should be getting,
with no slave attached). I would get FF for the second byte read, but
not the first.

So, work in progress.

This code also has a signon "signature" that the host checks when it
connects.