A more efficient TWI chat over USB
Still generic, but much more efficient for moving memory around.

The only thing making this not totally generic, is that it assumes that the
first byte sent to the slave (after the address) is a _command byte_; this and
the slave address (7 bits, *not* shifted left) are sent in the SETUP packet.

Still to do:

* pay attention to ACK bits while sending slave address and command

* add support for sending just a start bit, slave+write, and stop bit, and
  recording the ACK bit, as a way of pinging the target for readiness.
  Currently when it executes code, it first sets a bit so that it NACKs its
  slave address. When it returns from execution, it resumes ACKing its
  address.