Standardised ld/st in asm
But there are now prefix and postfix versions. Postfix keeps the
indexing operator nearer to the ld or st (it follows the op - hence the
name) and reads nicely.

Prefix is maybe a bit more like the "factory" assembler.

But both are quirky.

ld and st to data space (what Atmel calls lds and sts) I've also given
those names, even though I don't like them. My approach (following the
address with ")" or "$") didn't parse well, and the code in the asm was
too complicated. Pulling out lds and sts as the special cases that they
are (alone among the ld/st instructions in requiring a following offset
word) made things simpler and better.

I've updated the only AVR code I've written yet - the TWI chat - to work
with these changes.