AVR TWI chat code now callable as a subroutine
It saves the registers it uses. And it has a new fancy go command that
allows the host to set _all_ the registers - including X and Z (which
are used by the chat code) to arbitrary values before execution.

There is also a Bye command to return to the caller.

This all means that any code can call into chat to start an interactive
debug session, to look at memory or registers or whatever. Should make
debugging really nice.