Beginnings of programming PIC18 over USB (not working)
So I started to write this, and then after a careful tenth re-reading of the
datasheet about programming the PIC18F14K50, I realised I wasn't going to able
to use the JS16 as a programmer, at least as I've got it wired (running on 5v)
because the PGC and PGD pins on the 14K50 are shared with the D+ and D- USB
pins, and are speced only for USB's 3.3v logic levels.

So, back to using a 3v part to talk to the PIC - in this case a QG8 over
serial. So I'm committing this, but temporarily shelving it. It's way
incomplete anyway - I just copied over the avrprog-usb code and started
hacking on it, but didn't get very far before realising that it wasn't going
to work out.