Successfully read out PIC18 device id and configuration registers!
And because I can execute arbitrary code on the chip, it should be
possible to use the INDF registers to look around at the RAM and all the
SFRs too.

Cool!

Sadly I had to do this over serial and not USB, since I have voltage
issues with having the JS16 (5v) talk to 14K50 (3.3v). Maybe if I can
get USB code running on the 14K50 (!!) I can use it to program _other_
PIC parts.