Improved support for AVR image files, and better flash support
Now eeprom, flash, and fuses can be read from and written to the chip,
read from and written to an AVR image file, and verified (between the
in-memory versions and what's on the chip).

In the course of writing this code I discovered a bug in my treatment of
reading the eeprom using avr.BulkRead - I was using a _word_ address
(required when reading and programming the flash) instead of a byte
address - which is used to read and write the eeprom, even in page mode
(which we currently don't, on purpose, support, since it's not supported
on the 8515. Of course we currently use RDY/nBUSY to determine if a
flash or eeprom action has been completed, and these aren't supported on
8515 either...)