Much improved AVR interrupt vector support
Now by default target/AVR/build leaves room for a full set of vectors,
and it's possible, using the meta word "default-handler" (instead of
"handler") to default all the unset vectors to point to a particular
entry point - usually the same as the RESET vector.

The chat-twi code has been modified to use this new code, and the device
equates files have been updated so that they "call out" the last vector
that a particular chip defines.