Target serial port handling is now more flexible
Instead of having a hard-wired character device file called out in
lib/terminal.mu4 - which means maintaining a local change if your setup
is different from the sources - the serial/terminal code now expects to
see a symlink in muforth/ that points to the device that you want to
use.

The configure script tries (very simply) to guess, and to create a link
if none exists. It won't clobber an existing link though.

Also, the file opening has been improved. When doing a file search along
the search path, instead of of testing whether a file is readable (which
didn't work for symlinks!) it instead just tests whether it is
stat-able. Readability will be determined by the OS, and an obvious
error will result if permissions aren't sufficient.