Added "always..then" to the AVR asm
Since several tests in AVR assembler result in skipping (or not) the
following instruction, rather than compiling a conditional jump, I
figure it would be useful to have something like the assmebler's "if"
than instead compiles a forward (unconditional) jump, to the following
"then". Then a test-and-skip instruction will either skip - into the
conditional part - or fall thru - and jump around the conditional part.