Much better handling of export of git commit to muforth
Before it was a muddle of C, which only had the current, correct commit
if you had recently re-made muforth.

Now any git action that changes HEAD (clone, checkout, commit) will
write a commit.mu4 with the correct current version. This is now read by
startup and included in the banner.

The configure script notices if you're not using git, and generates a
"dummy" commit.mu4 script with a 0-len string. This prints no sha1 on
the banner.