Allow use of getcwd() when calculating file paths
By defining WITH_CURRENT_DIR duriing compilation, the behaviour of file
opening can be changed to allow the substitution of prefix "." with the
current working directory.

This is *not* the default. The code is ugly, and it's much simpler to
always have (relative) paths be rooted in the muforth build directory.