Changed how AVR image metadata is compiled
Now for each line of metadata about the image, a word in .meta. is
required; its _name_ is the "key" (to the left of the colon); its
execution creates (by compiling strings into the image) the value.

While this saves a few words in the host image, it's more complicated
and I'm not sure it's an improvement...