Target image metadata (AVR) now uses create/does
Every other way of doing this seemed error-prone. I liked how it read
when I used based-on inline, but it needed to be the first word
compiled; anywhere else and bad things would happen.

Doing it with create/does means that the embedded IP pointer does the
work of the embedded based-on, but there is no way _not_ to have it be
the first word, so there is less possibility for error.