muFORTH is 64 bit, everywhere, all the time!!
This commit re-implements something I tried in commit
69c30640c3078f751dfd5f39c3f5d4ccd3232598.

The basic idea: whether on a 32-bit machine or a 64-bit machine,
everything visible to the user (in Forth) is 64 bits wide. Cells, values
on the stacks, addresses, colon word bodies. Everything. @ and ! act on
64-bit values.

This seems like the only simple and uniform solution to the following
problem: I would love for muFORTH to be 64-bit on 64-bit machines, but I
don't want the experience to be different for 32-bit users. My earlier
compromise - of making values on the stacks, and all _computations_ - be
64 bits wide was great, but it had two problems (at least):

- storing and fetching those values dropped the high 32 bits

- forcing compilation of 32bit addresses was suboptimal (not allowing
  64-bit users to benefit from their hardware) and possibly buggy (eg,
  it didn't work correctly on FreeBSD/amd64 - the muforth executable
  would fail very early with a seg fault).

This approach is not without its problems, but they are very modest.
Actually there is really only one downside: memory usage. And the fact
is that doubling the size of the dict (for a basic, just-fired-up
muforth with nothing added) from 23k (for a pure 32-bit version) to 46k
or so (for a pure 64-bit version) is just not a big deal, considering
the hardware that muforth runs on: desktop computers with gigs of
memory. Even my 12-year-old PentiumPro laptop has 192MB of memory -
plenty for muforth's hoggy 8M heap, which is full of zeroes. And mostly
empty. Loading a "development system" for an 8-bit micro (AVR or HC08,
eg) takes another 60k or so.

Anyway, the point is this: in this new version, on a 32-bit machine, all
addresses are _stored_ as 64-bit values, but the high 32-bits are always
zero, and are actually ignored during execution: the C execution engine
does not load or store them, so there is no speed penalty. But
everything is twice as big.

But, who cares? It's just not a big deal, and I'm tired of
re-considering this idea over and over, of plowing and re-plowing this
field. I've planted it, and the seeds each take 64 bits. Enough.

One last comment: On 32-bit builds, machine addresses are often pushed
onto the stacks or stored in dict cells, both of which are now always 64
bits wide. So the question: do we zero- or sign-extend? For some reason,
when I went down similar paths before, I had to sign-extend addresses or
the system broke. This time around, since any data structure in memory
that holds a 32-bit pointer is going to have a high word of zeroes, it
seemed to make sense to zero-extend. It seems to work. If things break,
I'll re-assess this.

The definition that determines the behaviour of widening addresses is
the type  addr  , defined in muforth.h. It is currently typedef'd to be a
uintptr_t.

Enjoy!