Teasing apart USB support for BSD
FreeBSD and NetBSD used to share a USB implementation. Now the two are
different, in some subtle and not-so-subtle ways.

This commit is the first of several to modularise support for the two
OSes. I can successfully do control requests on both. I still need to
write the device "enumeration" code - to find a device based on its USB
vid and pid.