Added bulk endpoint support to Linux USB code!
And renamed a few things.

usb-find-interface is back to being usb-find-device, which also now
claims interface 0. (On OSX there are both usb-find-device and
usb-find-interface. This needs to be changed.)

usb-request is now usb-control (as in control transfer request).
usb-read-pipe is now just usb-read. Ditto usb-write.

I can connect to the STLINK and read out the "coreid". I'm still not
sure what this is... the IDCODE register of the SW-DP?

Anyway. Onward and upward!