Partly working memory access debug commands
Lots of refactoring of the cmd buffer code too. It was "neat" in a way
the way it _was_ but it was too inflexible. I'm not sure I'm anywhere
close to having a good factoring, and the mem access words fail oddly.

I tried to compare the output from st-flash (of reading the first 256
bytes of flash) to an image created by doing 4 64-byte reads (the
maximum bulk transfer I can do, given the endpt size) to pad, then
writing that out to a file, then doing two hexdumps side by side.

It seemed like my buffer was shifted, and the beginning was wrong. It
had good values towards the end, though. Not awake enough to debug this
right now. ;-)