Beginnings of a crib sheet for 16-bit Thumb instructions
I'm focussing initially on support for ARMv6-M architecture - which the
Cortex-M0 and -M0+ processors adhere to.

Because it's a strict subset of the v7-M (Cortex-M3/M4), I figured that
starting here made sense, as code and support for v6-M will run
unchanged on v7-M devices. (The reverse is definitely not true.)