ARM memory dump works
But I think there is a better way. The Cortex-M memory map is by nature
very simple. What about automatically mapping any address in the
(target) range 0000_0000 - 1fff_ffff to the flash image (wrapping as
necessary to make it fit) and mapping 2000_0000 - 3fff_ffff (wrapping)
to RAM?

Then it won't be necessary to switch images and get confused by being in
the wrong one.