Be careful when loading files to wait to reset zloading and line
If there is an error opening or reading a file, the message used to say
the wrong file name. It used to mention the _loaded_ file as the location
of the problem, rather than the _loading_ file.

I was resetting zloading and line before the file was successfully
opened. Now I wait until the last moment.