Removed h and w subcommands in v6-m du; refactored disasm a bit
In order to fix the terrible hack I introduced to make the ldr pc code
show the address it would load from and the value it would load, I
removed the two different cell widths ... which I had introduced mostly
to make by-hand disassembly easier - by showing halfword (16-bit) cells
instead of 32-bit cells.

Because there is now a (mostly complete) 16-bit Thumb disassembler,
dumping halfwords is much less useful, and it complicated both the code
and the UI.

I added a deferred word |@ (just for the v6-m target) to fetch a word
from an arbitrary address - rather than sequentially - and added an
implementation of it to the v6-m/memory code (image-@) and to the
v6-m/interact code (chat-@).