Made reading SREC files more flexible
Converting from a target address (ie, the address in an S record) to a
host address (ie, where to store it in memory in muforth) is now a
deferred word.

The default behaviour is to wrap the target address modulo 256Ki and add
pad. This way it's easy to use 'du' to test the behaviour.

A more "proper" usage would be to use 'image+' (for some target) and
store it into the memory image for a target chip.