Much cleaner setup for du/dis; ability for keys to save/restore mode
Now, when using c (call), v (vector), and r (return) keys when disasming
or dumping, not only the _address_ but the _mode_ is saved and restored.
So it's possible to go from dumping to disasming (this is what v does)
and back again.

It's even possible to mix and match dumping code on the host with code
on the target, though there is currently no way to easily make the jump
between them...

Overall, setting up for dumping and disasming is much cleaner. It's all
encapsulated in a create/does word.