Lots of small changes; still not a working asm though
Encodings are now explicit in the definitions instead of being
calculated; this way it's much easier to see where the instructions fit
into the encoding.

Several instructions simply work - push and pop eg - but there are lots
of pieces missing still.

Esp how to deal with the suffix S question...