Simplified creation of dict chains
Instead of creating two dict entries - a hidden one that acts like a
dict head, and a constant that points to it - create a does word that
simply pushes the address its body, which is a name entry structure. The
name is "muchain", to identify it (when dumping the dict, eg) but the
name is hidden (though it needn't be: find first dereferences the link
before beginning its search).