Added forth code to smudge and unsmudge colon words
It works, but I don't like it. Since I'm _toggling_ the hidden bit, it's
a bit fragile - one has to be careful not to re-smudge already unsmudged
words.

I'm thinking of abandoning the whole thing entirely, or doing it by
not linking colon words onto  current  until the word is complete.