First success with Freescale Freedom board!!
After adding code to usb-darwin.c to talk to HIDs (the Freedom
enumerates as a HID so that it can be used without drivers on Windows),
I was able to talk to the CMSIS-DAP firmware that I installed (which
sort of comes with the board).

I turned the connect LED on and off! Which means that communication
works, and the board is accepting my commands. Hooray!