Streamlined DAP.Transfer setup code
That is, the stuff inside <xfer and xfer>.

Now, not only is the SELECT register set only when needed (only when it
differs from the last value set), but the code takes care of counting
the "transfers" (DPACC or APACC) in the USB packet that we're writing.

This way the "client" code needn't be concerned - and in fact couldn't
get the number right anyway, since it doesn't know if a SELECT was
necessary or not (and each SELECT is an additional transfer).