Quick and dirty connection to Freedom board; can run du and dis!
Got memory dumping working, with a quick hack cobbled from the stlink
code. The choice of "operators" for the CMSIS-DAP interface is
suboptimal, but as an existence proof it works nicely. Right now "spy"
is on, so there is a lot of "noise", but it's possible to dump and
disassemble the contents of the chip.

Also added files to define the freedom board and its chip.