read and write byte/hword/word over cmsis-dap
Needed to get byte-laning working!

Didn't realise until now that reads and writes smaller than a word are
not shifted. The data is expected to be in (writes) or written to
(reads) the correct lane(s) of AP.DRW.

Another curious thing: reading flash or ram (which is to say: anything
connected directly to AHB) always reads the full word. But reading from
peripherals is more selective!