Removed (new) (create) (constant) (comment)
These words were all used intended to be flexible "plumbing" for a
second word (same name but without the parens) that first calls  token
and then calls its parenthetical self.

The whole point was to make it easy to write meta-compilers (and other
tricky code!) that re-uses tokens. But instead it just proliferates
complexity, because now you need two versions of a bunch of words.

Instead, we now have the word  untoken  which simply backs up the text
input pointer  first  to point at the beginning of the last token
parsed. It works great.

The ARM meta compiler is the only code that, to my knowledge, used one
of these parentheticals. So it's currently broken, but should be easy to
fix. All "modern" code - including HC08, AVR, and v6 (Cortex) ARM code
loads fine.