Added initial v7-m instruction encodings (32 bit thumb2)
These are mostly for load/store, because the encodings are really nuts.

I've added a bunch of instructions to the v7-m disassembler already that
are *not* reflected in the Instruction_encodings text file (yet).

But I really needed to put the load/store stuff down in a format that I
could parse visually so I could understand the encoding(s).