Reworked AVR chat-twi code
Not only is it now structured more like the HC08 code, with a register
frame on the stack, but I've added (optional) code to write the flash
via software - mostly to get an idea of how to write a bootloader.

Note: this is untested so far! Wanted to check it in to prevent
disastrous loss. ;-)