Reworked AVR tools for bootloader/app distinction
Now there are two regions in flash, one for the application (app) and
one for the bootloader (boot). (Using the bootloader space is optional.)

Flash programming via SPI was improved, and now programs the two spaces
independently. Also, a bug with programming fuses was fixed: previously
I wasn't doing a busy-wait to ensure completion.

Each device now "knows" how big (with BOOTSZ = 00, the maximum) its
bootloader space is.

I added self-programming support to chat-twi and tested it out. I found
that any Page Erase or Page Write command locks out the application
part of flash (the "no read while write" region) until the unlock
command (which I'm calling zclear for now, because it also clears the
page buffer) is executed.

But it needn't be done for each page. A bunch of pages could be erased
and reprogrammed, and one unlock done at the very end.