Support for remote execution, register display (AVR)
Since changing the way that the chat code is entered (now using a
partial "exception frame" containing registers 24-31 (ie, T, X, Y, and
Z) and SREG) it hasn't been possible to remotely execute code. That is
now working again, including support for pushing the top 8 stack entries
to the target, executing, and popping them back again.

The main new addition with this code is a register display. When
beginning a chat session and after any remote execution, the SREG, and
T, X, Y, Z, SP (hardware stack - Y is the forth data stack pointer), and
PC are displayed.