Require 3 op forms for many instructions (ARM asm and disasm)
For ARM v6-m it's not strictly necessary, but once we have the v7-m
assembler online it's going to be important to have consistent parsing
and display.

One problem is that a postfix assembler (like this one) can't tell how
many registers have been specified, unlike a normal prefix textual
assembler. So we can't use the same heuristics that ARM suggest in their
architecture manuals.

Instead, we basically require the user to specify the (often redundant)
Rd, even when it is the same as Rn, which it often is for the 16-bit
encodings.

And on diassembly, the 3 reg forms are shown when it makes sense to.