CMSIS-DAP: Tried (and failed) to make bulk reads/writes faster
Long story, but it's not that interesting. I was appalled by how long it
took to read the KL25's 128 KiB flash (12s) so I endeavored to make it
faster by writeing and then reading multiple 64 byte HID buffers...

It worked, but it was barely faster - taking 6.6s instead of 12s.

The code is here but it's commented out.

And it doesn't _entirely_ work...