Removed claiming of interface #0 on USB device open
The TI Launchpad boards have a debug interface with several USB
interfaces. The first two are for a debug serial port; interface #2 has
the bulk endpoints for talking to the debug stub. By separating out
opening the device and claiming an interface, it becomes possible to
talk to these boards.