Refactored digit? and fixed radix display code
I wanted to have  digit?  without the radix checking, so I created
digit>  and now  digit?  calls that.

While testing I noticed that nonstandard radices seem to keep pushing
themselves... The radix display code (part of the prompt) was fetching a
second copy of the radix and leaving it on the stack!