Beginnings of debug capability for Tiva/Stellaris Launchpad!
I can read and write memory and registers, and start and stop the
processor!

I don't exactly understand how the acknowledge/status process works
yet... and I just managed to get the firmware into a state that it
nacked everything I sent it...

So, some work to do. But pretty sweet so far!