ICDI: added smart-recv - for dealing with short bulk USB reads
Under VirtualBox (hosted on Windows) Linux seems to do a poor job of
reading from the bulk USB endpoints of the ICDI. In particular, shorter
buffers than expected are returned.

So, to deal with this, I've written "smart-recv" which keeps reading
until the last read ends with "#xx" - the packet end char and the
checksum.