Improved search for USB devices (again)
On some systems USB devices show up in _both_

  /proc/bus/usb

and

  /dev/bus/usb.

/proc/bus/usb is provided by usbfs (which has been, as of some 3.x Linux
kernel version, removed from the kernel) and has the feature (or not,
depending on your perspective) that all USB devices inherit the same
ownership and permissions. This was a simple and reliable system and
should not have been replaced!

However!!

Most modern Linux systems now run udevd, and it populates /dev/bus/usb
dynamically with USB devices as they are added and removed. However,
like all udevd devices, each USB device needs to be given a permission
and ownership. This is annoying and complicated, but I tried to simplify
it in a previous commit that generates udev rules for devices that work
with muforth.

Some systems (NixOS is one) retain /proc/sys/usb but also have
/dev/bus/usb. In this case, it's better to check _only_ /dev/bus/usb, so
this is what we do.