Failed attempt to trap bad reads (stellaris)
Reading a GPIO pin from the "AHB aperture" - eg, 4005_d000 for Port F -
causes the chip to go into LOCKUP.

Once I was able to see that it had gone into the Hard Fault handler,
which is helpfully a loop, so the PC remains the same once entered.

I put a vector table in RAM, pointed VTOR to it, and crossed my fingers,
but I was unable to get the chip to fault and enter my handler (which is
a single bkpt instruction).

Along the way I improved the "runwait" word - now it checks the DHCSR
for either HALT or LOCKUP state. The chip seems to enter one or the
other as soon I read the register. I have a loop to read it 8 times
before giving up, but the loop never loops!