Added editline (command line history) support!
Not yet loaded automagically into startup, but it's there and it works!

At first I wanted to have support for really long command lines (wider
than the screen, eg) - but it's soooo much easier to limit the user to
the width of their terminal instead - so that's what I did.

It's about 100 lines of Forth to add command line history (and editing)!
Pretty sweet.