Basic USB support added for DragonFly
And NetBSD/DragonFly/FreeBSD (before version 8) support _fixed_ to align
better with current practice. ;-)

Of course, there is work to do. There is no support for HIDs (and this
means CMSIS-DAP), or regular bulk endpoints (so no ST-LINK).

Getting those to work means thinking about the architecture of this
whole business, and may mean digging into the OSX support (ugh).