Brought cf/cf.mu4 back to life
Haven't touched cf/cf2.mu4.

This was fun to use with a big console font. On Linux, try this:

  $ setfont ter-132b

This will set a 16x32 pixel font that is nice for demos and such,
because it's very easy to see the screen. Unfortunately, even with my
1024x768 framebuffer console, since I only get 64x24 chars with this
font, muforth needs to be tweaked in a couple of places to not display
lines longer than 64 characters.

Other than that, it's pretty fun.