Minor refactoring of number output, including field (.r)
I was thinking about a right-to-left stack dump that converts multiple
numbers between  <#  and  #> . In order to make this work nicely, I had
to refactor the internals of a few words (most prominently  (.)  ) and
discovered that the definitions of  (.r)  and  (u.r)  made _no_ sense -
the blanks were not included in the converted number - so I fixed them
too.

One last thing: adding separators to numbers no longer depends on the
number being the only thing in the pad being converted. Instead of
checking  held  (the entire contents of number area) it checks how many
since the last fence; and it sets the fence before converting each
number. Pretty simple.