Revert "Minor refactoring of number output, including field (.r)"
This reverts commit 92955fa673c66e619b6c6b1362f9b32b60cdcf2f.

I'm not hoppy with how I'm doing number output - esp for strings that
are composites of multiples strings and numbers, such as rendered dates.
There has to be a simpler way.

Since I have a few ideas but I don't want to change this right now, I'm
backing out this change.

There are a few good ideas in here though; so I need to remember to take
a look at it when I revisit this problem!