Renamed h to 'h (target/AVR/memory)
This is all about the meta-compiler memory image for the AVR target.

'h is a variable that holds the address of a pair of pointers - to the
currently unallocated byte ("here") and to the origin, resp - that
together make up a region. So 'h preserve will preserve the region that
we are in.

We also define

: h  'h @ ;

h now acts like the "normal" variable that holds the current dictionary
pointer. IE, (meta) here is defined thus:

: here  h @ ;

So "h preserve" will remember where we are in the current space
(region).