Refactored AVR asm to better export jmp range checks
It's still a bit clunky - I really need to pass a _continuation_ to the
word that does the range checking; the continuation would be called if
the range check failed. For a normal asm word it would print an error.
For the meta-compiler it would instead compile a long (absolute) jmp or
call.

This would be cleaner. Right now it is a bit of a hack that requires
- the meta-compiler case - calculating the offset twice: once to check
it and again to compile the jmp/call.