AVR chat-twi expects app-init and app-task to exist
Well, sort of. Right now it _always_ calls a hook for each, but in the
future it needs to check if there is anything defined at those hooks. If
the app space has been erased we don't want to be trying to call into
it!

So this is a WIP of sorts.