If dumping memory while chatting, use -valid
Because we cache memory reads (in the dump code) in the case where we're
reading a line of memory from the image, printing it, and then doing the
same with a line of memory from the target, we'll get the cached image
values the second time unless we invalidate our dump cache.

So, in that one case, we do this.

Note that this would seem to be necessary much more often - say, when
switching from du (dump flash) to da (dump data). But the du-mode words
have _built-in_ a cache-inval operation! So a mode switch automatically
does this.