Added support for dumping EEPROM (image and chat)
While in there (target/AVR/memory, esp) decided to improve the treatment
of the data space. Now there is an actual data space image; it's
initialized with values that suggest "dead as a dodo" if it is
accidentally read.

Also, if  ,  or  c,  is executed while in ram image, a dummy image takes
the hit, rather than the previously current image - which could have
been eeprom-image or flash-image. This way things are a bit more robust
against accidental clobbering.