Check for "app present" flag before calling into application
Before I was checking the actual hook call sites, to see if the first
byte had been programmed. For various reasons - mostly having to do with
IDIOTIC architecture of the AVR - this didn't work. But even if I hadn't
run up against the different handling of _addresses into program space_
between the LPM and ICALL/IJMP instructions - the IDIOTIC!! design - I
would have had other problems.

What if I was in the middlle of programming a new version of the app,
using the bootloader. At some point I might have (probably _would_ have)
programmed the first byte of a the jump in one of the application hooks
_but not the whole instruction_. The first byte is programmed, so I
jump... off into space.

A much better, simpler design is to put a single byte at the end of the
application space. After completely programming the application, program
this one byte, and the application will start up.

I've decided to move the signature here as well. So the "app present"
byte ends up being the last byte of the signature space.