Tried to improve names of SPI and TWI plumbing files
Also changed a few of the "loading" remarks to better describe what's
happening.

It's a tricky business. There is code in target/HC08/ to program
firmware onto my JS16 board so I can program AVR parts over SPI. Naming
the pieces so it's clear what they are is *not* easy.