Added IICC synonyms to QE and JM parts for "generic" IIC support
The QE and JM subfamily added 10-bit slave address and "General call"
support to IIC. This required a second control register - IICC2 - so
IICC was rename IICC1.

However, IICC2 can be totally ignored - its RESET defaults are the same
as the older IIC hardware (eg, found in the QG parts) - and the bits in
IICC1 are exactly the same as IICC. So we make a synonym, to make
writing generic IIC code easier.