Bring target/HC08 up to speed with caching memory dumper
When I first wrote lib/du-cached, I decided to only switch the AVR and
ARM (v6/v7) code, and to leave the HC08 code. I wasn't actively doing
anything with the HC08 code at the time.

I'm not sure that I am now, but I'm _using_ it, and doing some
experiments, and I wanted it to be up to date.

The most complicated part of the change was to disentangle the
non-dumping uses of p@/p!/c*/c& etc. Since those all mean something very
different under the caching memory dumper, I thought it best to redefine
a special, simple, second set of pointer+byteread+bytewrite. (I've done
the same in the AVR code, but it is currently specialised for the flash
programming code. Maybe I should move it into target/AVR/memory and make
it more generic.)

I haven't tested the BDM code that resulted from this change, but
because it was a totally "mechanical" change it should be fine.