Working AVR IIC chat over serial
I tried this a long time ago and couldn't get it working. I gave up on
the S08's on-chip IIC interface, switched to an S08JS16 USB part - with
the advantage that I could now use USB - and bitbanged the IIC
interface.

I was able to establish comms with the AVR this way, and threw up my
hands over the S08's IIC.

I decided (for no good reason) to revisit this, and have figured out
working code - at least I think it works. It appears to give good data
when talking to the AVR, but for some reason the RXAK bit in IICS is
never correct, or at least never makes any sense. I've given up on it,
and now I ignore it.

Also, this code uses now obsolete "streaming" reads, such as zcell* and
zpcell*. These should be replaced with zread, zpread that read several
bytes into a buffer. This should be easy.