Streaming reads over IIC-over-serial work
I'm not sure I couldn't make the code a bit easier to read - esp for the
reads - and I may try. But this is all a bit of yak shaving anyway,
since the right approach is to put a bit more smarts into the
HC08-resident code, and not to "script" the IIC conversation over the
serial port.

But it's an interesting way to try to understand a protocol. I only wish
that the RXAK bit in the HCS08 would make sense...