frdm-kl25z test code
I have no idea what the status of this file is. It's obviously copied
from target/ARM/stellaris-test, which was used to execute a small amount
of code on a Stellaris Launchpad board.

The toplevel target/ARM/ directory is probably not the right place for
these files, but here they are... for now.