Moved gdbinit into src/ and renamed
I haven't run gdb on muforth for years. But I have kept an i386-specific
init file around. I wanted to get it out of the way.

I renamed it to .gdbinit-i386 and made a symlink

  .gdbinit -> .gdbinit-i386

So on x86 systems the right thing will happen, for those who want to run
gdb. On some other target, just change the symlink, or whatever you'd
like.