Reorder object files to make muforth faster?
It sounds crazy, but it seems to make a tiny bit of difference.

In general I'm impressed by how fast muforth is, considering that it's a
very simple system, with a fair amount of interpretive overhead - being
an indirect threaded code (ITC) system. My conjecture for why it ends up
still being pretty fast is that the executable C code is very small, and
the important bits of it likely fit into the cache of an x86 processor.

How long it'll stay there before the tasker swaps it out is another
matter, but in this case I think smallness and simplicity really win.

In any case, I'm happy about the result. Both that muforth is naturally
pretty fast, and that it's a tiny bit faster with this change. ;-)