Changed "or" back to "bis" in msp asm/disasm
In light of what I just learned about the _bis_ and _bic_ instructions -
namely, that they *don't* affect the condition codes - I think that
renaming _bis_ to _or_ was a mistake.

_and_ and _xor_ affect the CC as usual. _or_ suggests that it would act like
_and_ and _xor_. So calling the instructions _bis_ and _bic_ puts them
into another category. I think it's best to think of them as a species
of _move_ instruction, rather than bit-wise logical operators.